requirements.txt
djangorestframework
djangorestframework-simplejwt
pytest